Fixed KPP tolerance for non-UCX runs
The GCST identified that a reduction in the relative tolerance
was necessary to run the UCX in v10-01c without KPP non-convergence
errors, but it appears this adjustment is also necessary for
the non-UCX runs. 2.d-2 is a safe value, but may be too low. This
should be fully investigated later - 5.d-2 may be sufficient.

Signed-off-by: Sebastian D. Eastham <seastham@mit.edu>